Implement GET endpoints for properties

+++ src/main/java/com/engobytes/addressor/api/model/AutosearchSettingsApi.java
@@ -18,9 +18,9 @@ public class AutosearchSettingsApi {
     private int filteringLimit;
     private int photonApiLimit;
     private boolean useBoundary;
-    private float westBoundary;
-    private float northBoundary;
-    private float eastBoundary;
-    private float southBoundary;
+    private double westBoundary;
+    private double northBoundary;
+    private double eastBoundary;
+    private double southBoundary;
 
 }
